requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.16
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.0
psycopg2-binary==2.8.6
pytz==2024.1
requests==2.32.2
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.1
